% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.summary.R
\name{cc.summary}
\alias{cc.summary}
\title{Internal function to compute transformed summaries}
\usage{
cc.summary(result)
}
\description{
Internal function to compute transformed summaries
}
\keyword{internal}
