% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{get_vm_grid}
\alias{get_vm_grid}
\title{Helper to get or compute grid and memoize it}
\usage{
get_vm_grid(kappa, len = 2048L)
}
\description{
Helper to get or compute grid and memoize it
}
\keyword{internal}
