% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{lavm.link}
\alias{lavm.link}
\title{Create Link-Adjusted von Mises (LAvM) Link Object}
\usage{
lavm.link(type = "tan")
}
\arguments{
\item{type}{Character string defining the link function. Default is "tan".}
}
\value{
A list of class \code{lavm_link} containing the transform and inverse functions.
}
\description{
Create Link-Adjusted von Mises (LAvM) Link Object
}
