% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.family.R
\name{lavm.rename.inla.output}
\alias{lavm.rename.inla.output}
\title{Post-process INLA output to rename and exponentiate LAVM parameters}
\usage{
lavm.rename.inla.output(result, meta)
}
\description{
Post-process INLA output to rename and exponentiate LAVM parameters
}
\keyword{internal}
