% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{vm.spline.grid}
\alias{vm.spline.grid}
\title{Generate Spline Grid for Standard Von Mises (mu = 0)}
\usage{
vm.spline.grid(kappa, len = 2048L)
}
\arguments{
\item{kappa}{Numeric scalar. Concentration.}

\item{len}{Integer. Grid size.}
}
\description{
Generate Spline Grid for Standard Von Mises (mu = 0)
}
\keyword{internal}
