% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specfunc.R
\name{bessel_i}
\alias{bessel_i}
\title{Modified Bessel Function of the First Kind}
\usage{
bessel_i(x, nu, expon.scaled = FALSE)
}
\arguments{
\item{x}{numeric vector of values.}

\item{nu}{numeric scalar indicating the order of the Bessel function.}

\item{expon.scaled}{logical; if TRUE, the results are exponentially scaled
by \code{exp(-abs(x))}. Default is FALSE.}
}
\value{
A numeric vector of calculated Bessel values.
}
\description{
Evaluates the modified Bessel function of the first kind using high-precision
Chebyshev approximations.
}
