% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.model.R
\name{covariate}
\alias{covariate}
\title{Define Custom Covariate Settings}
\usage{
covariate(
  var,
  name = NULL,
  mean = NULL,
  sd = NULL,
  initial = NULL,
  fixed = FALSE,
  predictor = FALSE
)
}
\description{
Define Custom Covariate Settings
}
