% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlacc.R
\name{inlacc}
\alias{inlacc}
\title{Joint Circular Regression with INLA}
\usage{
inlacc(
  model,
  data,
  latent.index = NULL,
  LKJ.eta = 5,
  control.family = NULL,
  control.fixed = NULL,
  control.inla = list(cmin = 0, compute.initial.values = TRUE),
  metrics = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{model}{A list of 'likelihood' objects, or a single 'likelihood' object.}

\item{data}{A data.frame containing all covariates and responses.}

\item{latent.index}{Optional list of index() mappings.}

\item{LKJ.eta}{Positive scalar LKJ shape parameter used by every
`iidkd_LKJ` component. Defaults to `5`. Within each corresponding `f()`
term, `pc.prior.u` and `pc.prior.alpha` default to `1` and `0.5`,
respectively.}

\item{control.family}{Optional global list of control.family configurations.}

\item{control.fixed}{Optional list for fixed effects priors.}

\item{control.inla}{Optional control list passed to inla().}

\item{metrics}{Logical. Compute WAIC, DIC, CPO, etc.?}

\item{verbose}{Logical. Run INLA in verbose mode?}

\item{...}{Additional arguments passed to inla().}
}
\value{
A fitted INLA model.
}
\description{
Joint Circular Regression with INLA
}
\author{
Xiang Ye \email{xiang.ye@kaust.edu.sa}
}
