% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.family.R
\name{lavm.cloglike}
\alias{lavm.cloglike}
\title{Define the custom LAVM likelihood for INLA}
\usage{
lavm.cloglike(family.setting = NULL)
}
\arguments{
\item{family.setting}{A list of controls for the LAVM likelihood. Within
`hyper$kappa`, `log.initial` specifies the initial log-concentration and
`fixed = TRUE` fixes the log-concentration at that value.}
}
\value{
An INLA cloglike object.
}
\description{
Define the custom LAVM likelihood for INLA
}
