% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.summary.R
\name{print.inlacc_summary}
\alias{print.inlacc_summary}
\title{Print method for inlacc summaries}
\usage{
\method{print}{inlacc_summary}(x, decimal = NULL, metrics = TRUE, ...)
}
\description{
Print method for inlacc summaries
}
