% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.summary.R
\name{print.inlacc_time}
\alias{print.inlacc_time}
\title{Print method for inlacc time summary}
\usage{
\method{print}{inlacc_time}(x, ...)
}
\arguments{
\item{x}{An \code{inlacc_time} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for inlacc time summary
}
