% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.summary.R
\name{summary.inlacc}
\alias{summary.inlacc}
\title{Summary method for inlacc models}
\usage{
\method{summary}{inlacc}(object, decimal = 3L, ...)
}
\arguments{
\item{object}{A fitted \code{inlacc} model.}

\item{decimal}{Number of decimal places retained for printing.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An object of class \code{inlacc_summary}.
}
\description{
Summary method for inlacc models
}
